requirements.txt
tk
screeninfo
pystray
Pillow
plyer
pyinstaller
